refactor: final structure refactor
