+++ src/algorithms/ant.c
@@ -5,12 +5,12 @@
 #include <time.h>
 #include <math.h>
 #define NO_RECORDING
-#include "./libs/funcoes_cec_2015/cec15_test_func.h"
-#include "./libs/statistics.h"
-#include "./libs/types.h"
-#include "./libs/utils.h"
-#include "./libs/crossover.h"
-#include "./libs/log.h"
+#include "../libs/funcoes_cec_2015/cec15_test_func.h"
+#include "../libs/statistics.h"
+#include "../libs/types.h"
+#include "../libs/utils.h"
+#include "../libs/crossover.h"
+#include "../libs/log.h"
 
 #define DEBUG(x)
