+++ src/scripts/sh/coleta-info.sh
